test: add absolute seek test
move tests to different file

License: MIT
